\name{CTMC2list}
\alias{CTMC2list}
\title{
  Convert Between two representations of a Continuous Time Markov Chain.
}
\description{
  Convert Between two representations of a Continuous Time Markov Chain.
}
\usage{
CTMC2list(aCTMC)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{aCTMC}{
    CTMC obj
}
}
\details{
  Convert between two representations.
}
\value{
  
  return  a list.
}
