\name{doublebracket-methods} %%Note: can't use a name like "[[-methods"
%%or it won't be read.
\docType{methods}
\alias{[[,CTMC_many,ANY,ANY,ANY-method}
\alias{[[,CTMC_PO_many,ANY,ANY,ANY-method}
\title{Methods for Function \code{[[} in Package \pkg{DOBAD}}
\description{
  Methods for function \code{[[} in package \pkg{DOBAD}
}
\section{Methods}{
\describe{
\item{\code{signature(x = "CTMC_many", i = "ANY", j = "ANY", drop = "ANY")}}{
  Returns the indicated CTMC object.
}

\item{\code{signature(x = "CTMC_PO_many", i = "ANY", j = "ANY", drop = "ANY")}}{
  Returns the indicated CTMC_PO_1 object.
}
}}
\keyword{methods}
