\name{getBDMCsPOlist-methods}
\docType{methods}
\alias{getBDMCsPOlist}
\alias{getBDMCsPOlist-methods}
\alias{getBDMCsPOlist,CTMC_PO_many-method}
\title{ Methods for Function \code{getBDMCsPOlist} in Package \pkg{DOBAD}}
\description{
  Methods for function \code{getBDMCsPOlist} in package \pkg{DOBAD}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "CTMC_PO_many")}}{
  Return the list of CTMC_PO_1 objects corresponding to the CTMC_PO_many
  argument.
}
}}
\keyword{methods}
\keyword{accessor}
