\name{getMCstate}
\Rdversion{1.1}
\alias{getMCstate}
\title{
  Get the state of a CTMC at various times
}
\description{
  Returns the state of the CTMC at each of times in Ts.
}
\usage{
getMCstate(CTMC, Ts)
}
\arguments{
  \item{CTMC}{
    a CTMC.
}
  \item{Ts}{
    vector of times >0.
}
}
\details{

}
\value{
    Vector of length length(Ts) of integer states.

}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ }
\keyword{ }% __ONLY ONE__ keyword per line
