\name{getTimes}
\alias{getTimes}
\alias{getTimes,CTMC-method}
\alias{getTimes,BDMC-method}
\alias{getTimes,CTMC_PO_1-method}
\title{
  Get list of jump times.
}
\description{
  Object accessor. First jump  time is 0.
}
\usage{
getTimes(object)
}
\arguments{
  \item{object}{ A CTMC.  has a list of jump times.
}
}
\details{
  Gets list of jump times.
}
\value{
  numeric vector, positive values.
}

\author{
  Charles Doss
}
