\name{plot-methods}
\docType{methods}
\alias{plot-methods}
\alias{plot,CTMC,missing-method}
\alias{plot,CTMC_PO_1,missing-method}
\title{Plot CTMCs}
\description{
  Plotting for fully observed and partially observed Continuous Time
  Markov Chains.
}
\section{Methods}{
\describe{

\item{x = "CTMC", y = "missing"}{x is a fully observed CTMC.}

\item{x = "CTMC_PO_1", y = "missing"}{ x is discrete observations from a
   CTMC. }
}}
\keyword{methods}
