\name{getDataSummary.CTMC_PO_many}
\alias{getDataSummary.CTMC_PO_many}
\alias{getDataSummary}
\title{
  Calculate Some Summarizing Information for the Given Data
}
\description{
  Computes some summarizing statistics for a CTMC_PO_many object
  and returns them, possibly also saving them to a file.
}
\usage{
getDataSummary.CTMC_PO_many(dat, file = "dataSummary.rsav")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dat}{
    Discretely Observed BDI process.
}
  \item{file}{
    Filename to save to. 
}
}
\details{
  See the function definition for the variable names used.  Saving and
  loading to/from a file seemed like the simplest approach.
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
charles doss
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ }
\keyword{ }% __ONLY ONE__ keyword per line
