\name{list2CTMC}
\Rdversion{1.1}
\alias{list2CTMC}
\title{
  Convert a list representation of a CTMC to the class version
}
\description{
    Convert a list representation of a CTMC to the class version
}
\usage{
list2CTMC(aCTMC)
}
\arguments{
  \item{aCTMC}{
    A CTMC represented as a list.  Should have a "states", "times"
    vectors and a T numeric.
}
}
\details{
  
}
\value{
  Returns the same data but as an object of class CTMC.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ }
\keyword{ }% __ONLY ONE__ keyword per line
